rename posterior.py -> model.py and data.py -> stamp.py
