+++ forcepho/stamp.py
